// src/app/page/master/pajak/pajak.interface.ts
    import {EventEmitter} from '@angular/core';

    export interface pajak {
		kdProfile?: string;
		kdJenisPajak?: number;
		deskripsi?: string;
		kdDepartemen?: number;
		reportDisplay?: string;
		kodeExternal?: string;
		namaExternal?: string;
		tampilanLaporan?: string;
		noRec?: string;
		kode?:any;
		}

	